Add files via upload
These are two models: one for grayscale images and the other for RGB images. The test accuracy is 98% for both, while the validation accuracy is around 82%. We need to experiment with the models to increase the accuracy of the test and validation datasets. Overall, there doesn't seem to be any overfitting, and the models are performing well. I tried both Adam and SGD optimizers; SGD seems to provide better accuracy results.
Notes:

    Grayscale Model:
        The code is adapted from a GitHub repository originally designed for the MNIST dataset, which uses images of size 28x28. I transformed our dataset to 28x28. We could experiment with increasing the image size to 90x90, which is the original dataset size, to potentially improve the model's performance.

    RGB Model:
        For the RGB model, I used the code from Lab07 and transformed the images to 64x64 because CIFAR-10 images are 64x64, and the code relies on this size. We could also experiment with different model architectures to see whether RGB or grayscale performs better.

TODO:

    Create a second model with different hyperparameters and test it on both RGB and grayscale datasets.
    Experiment with different image sizes (e.g., 90x90 for grayscale) to see if it improves performance.
    Compare the performance of the RGB and grayscale models to determine which one performs better.